Issue #343: Simplified namespace. Moved template to qualified folder. Inlined CSS to avoid resolution issues. Avoided deprecated Freemarker code.
