--- desktop/ui/src/main/resources/documentationReference.css
@@ -1,63 +1,0 @@
-.body {
-	width: 900px;
-	margin: 0 auto;
-	font-family: 'Ubuntu';
-}
-
-.navbar {
-	position: relative;
-	min-height: 30px;
-	margin-bottom: 20px;
-	border: 1px solid;
-	background-color: #222;
-	border-color: #080808;
-	border-radius: 0;
-	background-image: url("images/menu/dc-logo-30.png");
-	background-repeat: no-repeat;
-}
-
-.navbar a {
-	text-decoration: none;
-}
-
-.navbar-brand {
-	padding: 3px 0px 0px 30px;
-	font-size: 20px;
-	color: white;
-	position: absolute;
-}
-
-.documentationBody .title {
-	padding-bottom: 25px;
-}
-
-.documentationBody .details {
-	padding-top: 3px;
-}
-
-.propertiesdetails {
-	text-align: left;
-	margin: 10px;
-}
-
-.propertiesdetails hr{
-	border-width: 1px; 
-/* 	border-color: #5591e2; */
-} 
-
-
-.propertiesdetails h4 {
-	margin-bottom: 10px;
-}
-
-.propertiesdetails .name {
-	color: #5591e2;
-}
-
-.propertiesdetails .required {
-	background-color: #f59b06;
-	color: white;
-	padding: 3px;
-	margin-left: 10px;
-	font-family: 'Arial'; 
-}
